Implement chat_info and channel_get_users
Actually those are stubs since via API you can't really get the member's
list (might be an idea to link that to a real tg-cli just to get the
real result, but it's something that can be done later).

Added a couple of fields previously missing on the conversion function.